Refactor worker objects and shared map
Introduce SharedMap (with to_json) and make file/result dataclasses inherit it. Move many per-item dataclasses (PageInfo, TemplateInfo, TemplateData, TemplateProcessingInfo, RenameInfo, ChartUpdateInfo) out of worker modules into their objects.py files to remove duplication and centralize state. Update routes and worker imports to use the new classes and to_json instead of to_dict. Affected files include files_service/objects.py, multiple admin_jobs_workers/*/objects.py and worker.py, and inject/extract routes. Refactor-only: no logic changes.
