+++ src/main_app/api_services/files_service/objects.py
@@ -3,49 +3,43 @@ from typing import Any
 
 
 @dataclass
-class FileLanguagesMap:
+class SharedMap:
+    error: str | None = None
+
+    def to_json(self) -> dict[str, Any]:
+        return asdict(self)
+
+
+@dataclass
+class FileLanguagesMap(SharedMap):
     """Result of extracting available SVG translation languages for a Commons file."""
 
-    error: str | None = None
     langs: list[str] | None = None
 
 
 @dataclass
-class FileInfo:
+class FileInfo(SharedMap):
     imageinfo: list[dict[str, Any]] | None = None
-    error: str | None = None
     exists: bool | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
 
 
 @dataclass
-class DownloadAndSaveData:
+class DownloadAndSaveData(SharedMap):
     result: str
-    error: str | None = None
     path: str | None = None
     size_bytes: int | None = None
 
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
-
 
 @dataclass
-class DownloadResult:
+class DownloadResult(SharedMap):
     success: bool = False
     size_bytes: int | None = None
     path: str | None = None
-    error: str | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        return asdict(self)
 
 
 @dataclass
-class UploadResult:
+class UploadResult(SharedMap):
     ok: bool | None
-    error: str | None = None
     error_details: str = ""
     msg: str | None = None
     result: dict[str, Any] | None = None
