+++ src/main_app/jobs_workers/admin_jobs_workers/add_lang_categories_to_owid_pages/worker.py
@@ -12,9 +12,6 @@ from __future__ import annotations
 
 import logging
 from collections.abc import Iterable
-from dataclasses import dataclass, field
-from datetime import datetime
-from typing import Any
 
 from mwclient.client import Site
 
@@ -23,52 +20,13 @@ from ....api_services.files_service.file_langs import get_file_languages
 from ....utils.wikitext.categories_utils import get_missing_categories_list
 from ...base_worker import BaseObjectsJobWorker
 from ...objects import JobsRunner
-from .objects import AddLangCategoriesWorkerObject
+from .objects import AddLangCategoriesWorkerObject, PageInfo
 from .utils import (
     build_category_names,
     extract_svg_file_name,
 )
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class PageInfo:
-    """Holds all state for a single OWID page being processed."""
-
-    page_title: str
-    svg_file: str | None = None
-    lang_codes: list[str] = field(default_factory=list)
-    categories_added: list[str] = field(default_factory=list)
-    timestamp: str = field(default_factory=lambda: datetime.now().isoformat())
-    status: str = "pending"
-    error: str | None = None
-    steps: dict[str, dict[str, Any]] = field(
-        default_factory=lambda: {
-            "load_page_text": {"result": None, "msg": ""},
-            "extract_file_name": {"result": None, "msg": ""},
-            "get_languages": {"result": None, "msg": ""},
-            "build_categories": {"result": None, "msg": ""},
-            "check_existing": {"result": None, "msg": ""},
-            "save_page": {"result": None, "msg": ""},
-        }
-    )
-
-    # Internal temporary state
-    _text: str | None = None
-    _categories: list[str] = field(default_factory=list)
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "page_title": self.page_title,
-            "svg_file": self.svg_file,
-            "lang_codes": self.lang_codes,
-            "categories_added": self.categories_added,
-            "timestamp": self.timestamp,
-            "status": self.status,
-            "error": self.error,
-            "steps": self.steps,
-        }
 
 
 class AddLangCategoriesWorker(BaseObjectsJobWorker):
