+++ src/main_app/jobs_workers/admin_jobs_workers/add_svglanguages_template/worker.py
@@ -5,9 +5,6 @@ Worker module for add_svglanguages_template.
 from __future__ import annotations
 
 import logging
-from dataclasses import dataclass, field
-from datetime import datetime
-from typing import Any
 
 from mwclient.client import Site
 
@@ -16,44 +13,10 @@ from ....db.models import TemplateRecord
 from ....db.services import TemplateService
 from ...base_worker import BaseObjectsJobWorker
 from ...objects import JobsRunner
-from .objects import AddSvgLanguagesWorkerObject
+from .objects import AddSvgLanguagesWorkerObject, TemplateInfo
 from .utils import RE_SVG_LANG, add_template_to_text, extract_svg_file_name
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass
-class TemplateInfo:
-    """Holds all state for a single template being processed."""
-
-    template_id: int
-    template_title: str
-    timestamp: str = field(default_factory=lambda: datetime.now().isoformat())
-    status: str = "pending"
-    error: str | None = None
-    steps: dict[str, dict[str, Any]] = field(
-        default_factory=lambda: {
-            "load_template_text": {"result": None, "msg": ""},
-            "generate_template_text": {"result": None, "msg": ""},
-            "add_template_text": {"result": None, "msg": ""},
-            "save_new_text": {"result": None, "msg": ""},
-        }
-    )
-
-    # Internal temporary state
-    _text: str | None = None
-    _template_text: str | None = None
-    _new_text: str | None = None
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "template_id": self.template_id,
-            "template_title": self.template_title,
-            "timestamp": self.timestamp,
-            "status": self.status,
-            "error": self.error,
-            "steps": self.steps,
-        }
 
 
 class AddSvgSVGLanguagesTemplate(BaseObjectsJobWorker):
