+++ src/main_app/jobs_workers/admin_jobs_workers/collect_templates_data/objects.py
@@ -63,6 +63,23 @@ class TemplateInfo:
 
 
 @dataclass
+class TemplateData:
+    id: int = 0
+    title: str = ""
+    main_file: str | None = None
+    last_world_file: str | None = None
+    last_world_year: int | None = None
+    slug: str = ""
+    source: str = ""
+
+    def __post_init__(self) -> None:
+        if self.main_file:
+            self.main_file = self.main_file.removeprefix("File:")
+        if self.last_world_file:
+            self.last_world_file = self.last_world_file.removeprefix("File:")
+
+
+@dataclass
 class CollectTemplatesDataWorkerObject(StandardAdminWorkerObject):
     pages_added: list[dict[str, Any]] = field(default_factory=list)
     pages_updated: list[dict[str, Any]] = field(default_factory=list)
